requirements.txt
aiogram==3.18.0
fastapi==0.115.8
uvicorn[standard]==0.34.0
jinja2==3.1.5
python-multipart==0.0.20
boto3==1.37.10
pydantic-settings==2.8.1
python-dotenv==1.0.1
httpx==0.28.1
itsdangerous==2.2.0
pytest==8.3.5
pytest-asyncio==0.25.3
